requirements.txt
flask
deepface
opencv-python
numpy
pillow
